fix: actions (#1)
* Update NEWS with issue #298

* Update for release

* Windows: fix size prefix specifiers of printf()

On Windows (mingw(-w64) and old Visual Studio, %z and %t are not
supported. Use the size prefix specifiers in msvcrt.dll. See :
https://docs.microsoft.com/en-us/cpp/c-runtime-library/format-specification-syntax-printf-and-wprintf-functions?view=vs-2019

* add documentation in libcompat.h

* Testing formatted README.md

* Center Repology distro list

* Created suite_create.3 man page

* Autotools build system configured to install man pages

* Modified the man page header to include library version number

* Created library man page

* Section 7 man pages now also installed by autotools

* Fixed Autoconf configuration files list spacing

* Man page installation via CMake now fully functional

* CMake automatic man page build date and version info complete

CMake now automatically configures the project man pages,
using a timestamp and the project version number from the
configure.ac file to ensure the user can clearly see the
library version number from which the documentation they
are reading originates.

* Automatic Autoconf Man Page Date/Versioning Complete

The only modifications required to replicate the automatic
man page version and date information previously
implemented with CMake was the addition of an m4 macro from
the GNU Autoconf Macro archive and the necessary calls from
configure.ac to utilize this new functionality.

* Fixed CMake man page installation bug

The man pages need to be prefixed with
${CMAKE_CURRENT_BINARY_DIR} in order for CMake to actually
look in the configuration directory as opposed to the
directory where they were originally located.

I'm not sure how I missed this bug the first go-around, but
since the man pages are optional, CMake silently moves on
if it can't find the man pages where it thinks they should
be.

This bug has now been fixed (hopefully).

* Removed Timestamps from Project Configuration Process

As discussed in the pull request (PR #310), the addition of
timestamps to the project configuration, and transitively
to the man pages, we sacrificed build reproducibility
without really getting much in return.

Therefore, this pull request removes the timestamp function
from the CMakeLists.txt file which contained it, as well as
the corresponding Autoconf macro file and the code that
called it from the Autotools build system.

Furthermore, the man pages' date section has been replaced
with an empty string, but the project version is still
configured via either CMake or Autotools onto the man pages
so that the reader is aware of the library version from
which they originated.

* Fixed wrong function replacement

The source file used to replace the strdup function was
different for the shared and static builds of the library.

* Created GitHub Bug Report Issue Template

I created a GitHub issue template specifically for bug
reports to help streamline the process of pinning down,
replicating, verifying, and fixing user-reported issues.

* Created GitHub Issues Configuration File

The config.yml at the root of the .github/ISSUE_TEMPLATE
directory controls the issue templates configuration for
the project directory.

Eventually, it would be nice to not have to accept blank
issues, since having an issue template for each kind of
issue helps speed up the process by allowing us to define
the necessary information we need up front.

While we create the issue templates, however, I'm leaving
blank issues enabled.

I also added the user mailing list address from the website
in case someone prefers to seek the wisdom therein.

* Removed User Mailing List From the Issues Template

As discussed in the pull request (#315), the user mailing
list has been removed to prevent the bifurcation of the
user base.

It is preferable to have all user questions, feature
requests, and bug reports filed here via the GitHub issues
feature.

Discussion: https://github.com/libcheck/check/pull/315#discussion_r515748364

* docs: fix simple typo, inproperly -> improperly

There is a small typo in tests/check_check_sub.c.

Should read `improperly` rather than `inproperly`.

* Allow use of deprecated add-path

The add-path command in Github Actions has
been deprecated:

https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/

This commit enables it use for a little longer, following
the example from:

https://github.com/actions/toolkit/issues/641

* check_mem_leaks: integrate check_mem_leaks into CMake

Commit 8699473b65011b50766d536222329eeb3953f1cc initialy introduced a
new test, check_mem_leaks, that can be used with Valgrind to find memory
errors.

Since then, test check_mem_leaks can be built with GNU Autotools scripts
but lacks integration into CMake scripts. This means that, when CMake is
used to build the project, check_mem_leaks will NOT be built.

This commit resolves this by adding support for building check_mem_leaks
into CMake scripts as well.

Signed-off-by: Hrvoje Varga <hrvoje.varga@gmail.com>

* Correct misspelling

* Adjust test suite for 106-bit long double precision

On PowerPC architectures (ppc, ppc64el, powerp) 'long double' has a
precision of 106-bit, compared to 80-bit precision on amd64.

This leads to the test_ck_assert_(float|double|ldouble)_eq_tol succeed
rather than fail as expected, cause 0.003-0.002 will be actually
slightly bigger than 0.001 and not slightly smaller.

Increase the change to the tolerance, so it will be on all architectures
smaller than the difference of ~0.001 and the unit tests will fail as
expected.

* Update how casks are installed for osx tests

The way to install casks for brew has changed, 
causing the tests to fail. Updating to the new
command syntax.

* Drop the remaining cvsignore files

The main .cvsignore file got deleted in 6f5da9a032 but that commit missed the
ones in the various directories.

* Fix pkgconfig file's libdir value

And make a similar fix for includedir while we're here.

Bug: https://bugs.gentoo.org/729642

* Ignore scan-build results for snprintf replacement

scan-build updated and is identifying issues with
the sprintf replacement:
https://github.com/libcheck/check/pull/327/checks?check_run_id=2135844732

This needs to be investigated and resolved. However,
it need not hold up other changes being submitted.

* Remove docs building test for OSX

The docs building test also builds on Linux,
and doing this also on OSX is unnecessary.
Building LaTeX should be the same on both
platforms. Running on both was unnecessary.

* Update CMake required from 3.9 to 3.13

https://github.com/libcheck/check/issues/325

    cmake_minimum_required(...) is version 3.9 required in top level CMakeLists.txt, In this file, the funcion add_link_options(...) is invoked at line 251 and it`s news from CMAKE version 3.13.

* Inform compiler (MinGW) that format strings to use are GNU/c99 style, not Windows

Co-authored-by: Branden Archer <b.m.archer4@gmail.com>
Co-authored-by: Vincent Torri <vincent.torri@gmail.com>
Co-authored-by: Jose Fernando Lopez Fernandez <jflopezfernandez@gmail.com>
Co-authored-by: Tim Gates <tim.gates@iress.com>
Co-authored-by: Hrvoje Varga <hrvoje.varga@gmail.com>
Co-authored-by: Christian Göttsche <cgzones@googlemail.com>
Co-authored-by: Peter Hutterer <peter.hutterer@who-t.net>
Co-authored-by: Matt Turner <mattst88@gmail.com>
Co-authored-by: Mikko Johannes Koivunalho <mikko.koivunalho@iki.fi>
Co-authored-by: bradh352 <brad@brad-house.com>